refactor: apply reviews
