+++ example/package.json
@@ -12,16 +12,16 @@
     "start": "ts-node --transpile-only -r tsconfig-paths/register src/index.ts"
   },
   "dependencies": {
-    "@nestjs/common": "7.0.3",
-    "@nestjs/core": "7.0.3",
-    "@nestjs/platform-express": "7.0.3",
+    "@nestjs/common": "7.0.7",
+    "@nestjs/core": "7.0.7",
+    "@nestjs/platform-express": "7.0.7",
     "reflect-metadata": "0.1.13",
-    "rxjs": "6.5.4",
-    "stripe": "8.32.0"
+    "rxjs": "6.5.5",
+    "stripe": "8.39.0"
   },
   "devDependencies": {
-    "@types/node": "13.9.0",
-    "ts-node": "8.6.2",
+    "@types/node": "13.11.0",
+    "ts-node": "8.8.1",
     "tsconfig-paths": "3.9.0",
     "typescript": "3.8.3"
   }
