+++ package.json
@@ -25,18 +25,18 @@
     "test": "jest"
   },
   "devDependencies": {
-    "@nestjs/common": "^7.0.3",
-    "@nestjs/core": "^7.0.3",
-    "@nestjs/testing": "^7.0.3",
-    "@types/jest": "25.1.4",
-    "@types/node": "13.9.0",
+    "@nestjs/common": "^7.0.7",
+    "@nestjs/core": "^7.0.7",
+    "@nestjs/testing": "^7.0.7",
+    "@types/jest": "25.2.1",
+    "@types/node": "13.11.0",
     "codecov": "3.6.5",
-    "jest": "25.1.0",
-    "prettier": "2.0.1",
+    "jest": "25.2.7",
+    "prettier": "2.0.2",
     "reflect-metadata": "^0.1.12",
     "rimraf": "3.0.2",
-    "rxjs": "^6.0.0",
-    "stripe": "^8.32.0",
+    "rxjs": "^6.5.5",
+    "stripe": "^8.39.0",
     "ts-jest": "25.2.1",
     "typescript": "3.8.3"
   },
