+++ src/constants/index.ts
@@ -3,4 +3,4 @@ import Stripe from 'stripe';
 export * from './stripeModuleOptions';
 export * from './stripeToken';
 
-export const apiVersion: Stripe.LatestApiVersion = '2020-03-02'
+export const apiVersion: Stripe.LatestApiVersion = '2020-03-02';
